feat: refactor leaflet map into another component and lazy load it
